pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "verge-browser"
version = "0.1.0"
description = "Browser sandbox service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.32,<1.0",
  "pydantic-settings>=2.6,<3.0",
  "httpx>=0.28,<1.0",
  "websockets>=14,<16",
  "python-multipart>=0.0.20,<1.0",
  "pyjwt>=2.10,<3.0",
  "orjson>=3.10,<4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.25,<1.0",
]

[tool.setuptools]
packages = []

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

